requirements.txt
aiogram>=3.0.0
python-dotenv>=0.19.0
aiohttp>=3.9.0
aiofiles>=23.2.1
pydantic>=2.4.1
typing-extensions>=4.7.0
magic-filter>=1.0.12
certifi>=2023.7.22
APScheduler>=3.10.1 